remove workflow version resource
