+++ dockstore-webservice/src/main/resources/swagger.yaml
@@ -53,7 +53,7 @@ paths:
     get:
       tags:
       - "aliases"
-      summary: "Retrieves a workflow version by alias."
+      summary: "Retrieves workflow version path information by alias."
       description: "Does not require authentication for published workflows, authentication\
         \ can be provided for restricted workflows"
       operationId: "getWorkflowVersionPathInfoByAlias"
@@ -69,7 +69,7 @@ paths:
         200:
           description: "successful operation"
           schema:
-            $ref: "#/definitions/WorkflowVersion"
+            $ref: "#/definitions/WorkflowVersionPathInfo"
       security:
       - BEARER: []
   /aliases/workflow-versions/{workflowVersionId}:
@@ -8202,6 +8202,13 @@ definitions:
         - "KUBERNETES"
         - "HELM"
     description: "This describes one workflow version associated with a workflow."
+  WorkflowVersionPathInfo:
+    type: "object"
+    properties:
+      fullWorkflowPath:
+        type: "string"
+      tagName:
+        type: "string"
 externalDocs:
   description: "Dockstore documentation"
   url: "https://www.dockstore.org/docs/getting-started"
